Implement a (hopefully) portable entropy collection suite
This is a one-file revamp of ottery's entropy collection
architecture, where we try to get at least one input from each of
several sources, avoiding inputs that we don't like if we already
have good ones, and so forth.

I am treating anything that comes from the OS via a file as separate
from stuff that comes from the OS via a syscall.  This is redundant
and stoopid, but what can it hurt?

I am also including RDRAND, because even if conjectures about an
RDRAND backdoor are true, there are probably some attackers who
can't exploit the (conjectured) backdoor.  I'm not marking RDRAND as
strong.

Taking a hint from the libressl crew, I'm adding a fallback rng for
everything but windows.  It looks at a bunch of proc files and
sysctl outputs and timers and whatnot. The list of sysctls and proc
files and stuff here is my own.  Perhaps I should try some of their
mmap trickery too.

I've taken ideas from the libressl implementation, the tor
implementation, and the libevent implementation.  All of the code is
new, though.

Based on a blog post by DJB [DJB], I'm trying to fetch more
questionable sources before stronger ones.  The threat model here
seems a little odd to me: I think it assumes an attacker with
partial control of the local host, who has read-only access to RAM
(or to some-but-not-all entropy sources), and who has read-write
access to others, and also it assumes that an attacker who can
observe and partially control your PRNG output is significantly
worse than one who can only observe it.  But DJB is much smarter
than I am, so I'm going to take the more obvious countermeasure.

[DJB] http://blog.cr.yp.to/20140205-entropy.html