Use minherit on rng contents
If we have MAP_INHERIT_ZERO like openbsd does, we can zero the whole
thing, and avoid getpid() calls.

If we only have MAP_INHERIT_NONE like other BSDs (and OSx) do, we can
ensure that we get a segfault if our getpid() checks ever fail.

We should eliminate the redundant getpid() check when we can.