Have an API for checking the status.
This is a divergence from the arc4random philosophy.  Me, I like a way
for a programmer to ask how good the entropy source is, and maybe to
exit gracefully in cases where we would otherwise abort.

Most programs should just not call this function.  Programs like tor
should call it, then exit if the value is not 2.  Some programs should
call it before even starting the library, and exit nicely if the value
is 0.